repl: output file for integration tests (#689)
