+++ kore/src/Kore/Repl/Data.hs
@@ -18,62 +18,41 @@ module Kore.Repl.Data
     , NodeState (..)
     , GraphProofStatus (..)
     , AliasDefinition (..), ReplAlias (..), AliasArgument(..), AliasError (..)
-    , getNodeState
     , InnerGraph
     , lensAxioms, lensClaims, lensClaim
     , lensGraphs, lensNode, lensStepper
     , lensLabels, lensOmit, lensUnifier
     , lensCommands, lensAliases, lensClaimIndex
-    , lensLogging
+    , lensLogging, lensLogger
+    , lensOutputFile
     , shouldStore
+    , commandSet
     , UnifierWithExplanation (..)
     , runUnifierWithExplanation
-    , emptyExecutionGraph
-    , getClaimByIndex, getAxiomByIndex, getAxiomOrClaimByIndex
-    , switchToProof
-    , getTargetNode, getInnerGraph, getExecutionGraph
-    , getConfigAt, getRuleFor, getLabels, setLabels
     , StepResult(..)
-    , runStepper, runStepper'
-    , runUnifier
-    , updateInnerGraph, updateExecutionGraph
-    , addOrUpdateAlias, findAlias, substituteAlias
     , LogType (..)
+    , ReplScript (..)
+    , ReplMode (..)
+    , OutputFile (..)
     ) where
 
 import           Control.Applicative
                  ( Alternative )
 import           Control.Concurrent.MVar
-import qualified Control.Lens as Lens hiding
-                 ( makeLenses )
 import qualified Control.Lens.TH.Rules as Lens
-import           Control.Monad
-                 ( join )
-import           Control.Monad.Error.Class
-                 ( MonadError )
-import qualified Control.Monad.Error.Class as Monad.Error
-import           Control.Monad.IO.Class
-                 ( liftIO )
-import           Control.Monad.State.Strict
-                 ( MonadState, get, modify )
 import           Control.Monad.Trans.Accum
                  ( AccumT, runAccumT )
 import qualified Control.Monad.Trans.Accum as Monad.Accum
 import qualified Control.Monad.Trans.Class as Monad.Trans
-import           Data.Bitraversable
-                 ( bisequence, bitraverse )
-import           Data.Coerce
-                 ( coerce )
 import qualified Data.Graph.Inductive.Graph as Graph
 import           Data.Graph.Inductive.PatriciaTree
                  ( Gr )
 import           Data.List.NonEmpty
                  ( NonEmpty (..) )
-import qualified Data.Map as Map
 import           Data.Map.Strict
                  ( Map )
 import           Data.Maybe
-                 ( fromMaybe, listToMaybe )
+                 ( fromMaybe )
 import           Data.Monoid
                  ( First (..) )
 import           Data.Sequence
@@ -81,21 +60,12 @@ import           Data.Sequence
 import           Data.Set
                  ( Set )
 import qualified Data.Set as Set
-import           Data.Text
-                 ( Text )
 import           Data.Text.Prettyprint.Doc
                  ( Doc )
 import qualified Data.Text.Prettyprint.Doc as Pretty
-import           GHC.Exts
-                 ( toList )
 import           Numeric.Natural
-import           System.IO
-                 ( Handle, IOMode (AppendMode), hClose, openFile )
 
-import           Kore.Internal.Conditional
-                 ( Conditional (..) )
-import           Kore.Internal.Predicate
-                 ( Predicate )
+import qualified Kore.Internal.Predicate as IPredicate
 import           Kore.Internal.TermLike
                  ( TermLike )
 import qualified Kore.Logger.Output as Logger
@@ -103,7 +73,7 @@ import           Kore.OnePath.StrategyPattern
 import           Kore.OnePath.Verification
                  ( Axiom (..), Claim )
 import           Kore.Step.Rule
-                 ( RewriteRule (..), RulePattern (..) )
+                 ( RewriteRule (..) )
 import           Kore.Step.Simplification.Data
                  ( MonadSimplify, Simplifier )
 import qualified Kore.Step.Strategy as Strategy
@@ -117,6 +87,19 @@ import           Kore.Unparser
                  ( unparse )
 import           SMT
                  ( MonadSMT )
+
+-- | Represents an optional file name which contains a sequence of
+-- repl commands.
+newtype ReplScript = ReplScript
+    { unReplScript :: Maybe FilePath
+    } deriving (Eq, Show)
+
+data ReplMode = Interactive | RunScript
+    deriving (Eq, Show)
+
+newtype OutputFile = OutputFile
+    { unOutputFile :: Maybe FilePath
+    } deriving (Eq, Show)
 
 newtype AxiomIndex = AxiomIndex
     { unAxiomIndex :: Int
@@ -381,7 +364,7 @@ data ReplState claim = ReplState
     , unifier
         :: TermLike Variable
         -> TermLike Variable
-        -> UnifierWithExplanation (Predicate Variable)
+        -> UnifierWithExplanation (IPredicate.Predicate Variable)
     -- ^ Unifier function, it is a partially applied 'unificationProcedure'
     --   where we discard the result since we are looking for unification
     --   failures
@@ -391,6 +374,7 @@ data ReplState claim = ReplState
     -- ^ Map of command aliases
     , logging :: (Logger.Severity, LogType)
     , logger  :: MVar (Logger.LogAction IO Logger.LogMessage)
+    , outputFile :: OutputFile
     }
 
 type Explanation = Doc ()
@@ -466,203 +450,6 @@ runUnifierWithExplanation (UnifierWithExplanation unifier) =
 
 Lens.makeLenses ''ReplState
 
--- | Creates a fresh execution graph for the given claim.
-emptyExecutionGraph :: Claim claim => claim -> ExecutionGraph
-emptyExecutionGraph =
-    Strategy.emptyExecutionGraph . extractConfig . RewriteRule . coerce
-  where
-    extractConfig
-        :: RewriteRule Variable
-        -> CommonStrategyPattern
-    extractConfig (RewriteRule RulePattern { left, requires }) =
-        RewritePattern $ Conditional left requires mempty
-
--- | Get nth claim from the claims list.
-getClaimByIndex
-    :: MonadState (ReplState claim) m
-    => Int
-    -- ^ index in the claims list
-    -> m (Maybe claim)
-getClaimByIndex index = Lens.preuse $ lensClaims . Lens.element index
-
--- | Get nth axiom from the axioms list.
-getAxiomByIndex
-    :: MonadState (ReplState claim) m
-    => Int
-    -- ^ index in the axioms list
-    -> m (Maybe Axiom)
-getAxiomByIndex index = Lens.preuse $ lensAxioms . Lens.element index
-
--- | Transforms an axiom or claim index into an axiom or claim if they could be
--- found.
-getAxiomOrClaimByIndex
-    :: MonadState (ReplState claim) m
-    => Either AxiomIndex ClaimIndex
-    -> m (Maybe (Either Axiom claim))
-getAxiomOrClaimByIndex =
-    fmap bisequence
-        . bitraverse
-            (getAxiomByIndex . coerce)
-            (getClaimByIndex . coerce)
-
--- | Update the currently selected claim to prove.
-switchToProof
-    :: MonadState (ReplState claim) m
-    => Claim claim
-    => claim
-    -> ClaimIndex
-    -> m ()
-switchToProof claim cindex =
-    modify (\st -> st
-        { claim = claim
-        , claimIndex = cindex
-        , node = ReplNode 0
-        })
-
--- | Get the internal representation of the execution graph.
-getInnerGraph
-    :: MonadState (ReplState claim) m
-    => Claim claim
-    => m InnerGraph
-getInnerGraph =
-    fmap Strategy.graph getExecutionGraph
-
--- | Get the current execution graph
-getExecutionGraph
-    :: MonadState (ReplState claim) m
-    => Claim claim
-    => m ExecutionGraph
-getExecutionGraph = do
-    ReplState { claimIndex, graphs, claim } <- get
-    let mgraph = Map.lookup claimIndex graphs
-    return $ maybe (emptyExecutionGraph claim) id mgraph
-
--- | Update the internal representation of the current execution graph.
-updateInnerGraph
-    :: MonadState (ReplState claim) m
-    => InnerGraph
-    -> m ()
-updateInnerGraph ig = do
-    ReplState { claimIndex, graphs } <- get
-    lensGraphs Lens..=
-        Map.adjust (updateInnerGraph0 ig) claimIndex graphs
-  where
-    updateInnerGraph0 :: InnerGraph -> ExecutionGraph -> ExecutionGraph
-    updateInnerGraph0 graph Strategy.ExecutionGraph { root } =
-        Strategy.ExecutionGraph { root, graph }
-
--- | Update the current execution graph.
-updateExecutionGraph
-    :: MonadState (ReplState claim) m
-    => ExecutionGraph
-    -> m ()
-updateExecutionGraph gph = do
-    ReplState { claimIndex, graphs } <- get
-    lensGraphs Lens..= Map.insert claimIndex gph graphs
-
--- | Get the node labels for the current claim.
-getLabels
-    :: MonadState (ReplState claim) m
-    => m (Map String ReplNode)
-getLabels = do
-    ReplState { claimIndex, labels } <- get
-    let mlabels = Map.lookup claimIndex labels
-    return $ maybe Map.empty id mlabels
-
--- | Update the node labels for the current claim.
-setLabels
-    :: MonadState (ReplState claim) m
-    => Map String ReplNode
-    -> m ()
-setLabels lbls = do
-    ReplState { claimIndex, labels } <- get
-    lensLabels Lens..= Map.insert claimIndex lbls labels
-
-
--- | Get selected node (or current node for 'Nothing') and validate that it's
--- part of the execution graph.
-getTargetNode
-    :: MonadState (ReplState claim) m
-    => Claim claim
-    => Maybe ReplNode
-    -- ^ node index
-    -> m (Maybe ReplNode)
-getTargetNode maybeNode = do
-    currentNode <- Lens.use lensNode
-    let node' = unReplNode $ maybe currentNode id maybeNode
-    graph <- getInnerGraph
-    if node' `elem` Graph.nodes graph
-       then pure . Just . ReplNode $ node'
-       else pure $ Nothing
-
--- | Get the configuration at selected node (or current node for 'Nothing').
-getConfigAt
-    :: MonadState (ReplState claim) m
-    => Claim claim
-    => Maybe ReplNode
-    -> m (Maybe (ReplNode, CommonStrategyPattern))
-getConfigAt maybeNode = do
-    node' <- getTargetNode maybeNode
-    case node' of
-        Nothing -> pure $ Nothing
-        Just n -> do
-            graph' <- getInnerGraph
-            pure $ Just (n, getLabel graph' (unReplNode n))
-  where
-    getLabel gr n = Graph.lab' . Graph.context gr $ n
-
--- | Get the rule used to reach selected node.
-getRuleFor
-    :: MonadState (ReplState claim) m
-    => Claim claim
-    => Maybe ReplNode
-    -- ^ node index
-    -> m (Maybe (RewriteRule Variable))
-getRuleFor maybeNode = do
-    targetNode <- getTargetNode maybeNode
-    graph' <- getInnerGraph
-    pure $ fmap unReplNode targetNode >>= getRewriteRule . Graph.inn graph'
-  where
-    getRewriteRule
-        :: forall a b
-        .  [(a, b, Seq (RewriteRule Variable))]
-        -> Maybe (RewriteRule Variable)
-    getRewriteRule =
-        listToMaybe
-        . join
-        . fmap (toList . third)
-
-    third :: forall a b c. (a, b, c) -> c
-    third (_, _, c) = c
-
--- | Lifting function that takes logging into account.
-liftSimplifierWithLogger
-    :: forall a t claim
-    .  MonadState (ReplState claim) (t Simplifier)
-    => Monad.Trans.MonadTrans t
-    => MVar (Logger.LogAction IO Logger.LogMessage)
-    -> Simplifier a
-    -> t Simplifier a
-liftSimplifierWithLogger mLogger simplifier = do
-   (severity, logType) <- logging <$> get
-   (textLogger, maybeHandle) <- logTypeToLogger logType
-   let logger = Logger.makeKoreLogger severity textLogger
-   _ <- Monad.Trans.lift . liftIO $ swapMVar mLogger logger
-   result <- Monad.Trans.lift simplifier
-   maybe (pure ()) (Monad.Trans.lift . liftIO . hClose) maybeHandle
-   pure result
-  where
-    logTypeToLogger
-        :: LogType
-        -> t Simplifier (Logger.LogAction IO Text, Maybe Handle)
-    logTypeToLogger =
-        \case
-            NoLogging   -> pure (mempty, Nothing)
-            LogToStdOut -> pure (Logger.logTextStdout, Nothing)
-            LogToFile file -> do
-                handle <- Monad.Trans.lift . liftIO $ openFile file AppendMode
-                pure (Logger.logTextHandle handle, Just handle)
-
 -- | Result after running one or multiple proof steps.
 data StepResult
     = NoResult
@@ -672,72 +459,6 @@ data StepResult
     | BranchResult [ReplNode]
     -- ^ configuration branched
     deriving (Show)
-
--- | Run a single step for the data in state
--- (claim, axioms, claims, current node and execution graph).
-runStepper
-    :: MonadState (ReplState claim) (m Simplifier)
-    => Monad.Trans.MonadTrans m
-    => Claim claim
-    => m Simplifier StepResult
-runStepper = do
-    ReplState { claims, axioms, node } <- get
-    (graph', res) <- runStepper' claims axioms node
-    updateExecutionGraph graph'
-    case res of
-        SingleResult nextNode -> do
-            lensNode Lens..= nextNode
-            pure res
-        _                     -> pure res
-
--- | Run a single step for the current claim with the selected claims, axioms
--- starting at the selected node.
-runStepper'
-    :: MonadState (ReplState claim) (m Simplifier)
-    => Monad.Trans.MonadTrans m
-    => Claim claim
-    => [claim]
-    -> [Axiom]
-    -> ReplNode
-    -> m Simplifier (ExecutionGraph, StepResult)
-runStepper' claims axioms node = do
-    ReplState { claim, stepper } <- get
-    mvar <- Lens.use lensLogger
-    gph <- getExecutionGraph
-    gr@Strategy.ExecutionGraph { graph = innerGraph } <-
-        liftSimplifierWithLogger mvar $ stepper claim claims axioms gph node
-    pure . (,) gr $ case Graph.suc innerGraph (unReplNode node) of
-        []       -> NoResult
-        [single] -> case getNodeState innerGraph single of
-                        Nothing -> NoResult
-                        Just (StuckNode, _) -> NoResult
-                        _ -> SingleResult . ReplNode $ single
-        nodes    -> BranchResult $ fmap ReplNode nodes
-
-runUnifier
-    :: MonadState (ReplState claim) (m Simplifier)
-    => Monad.Trans.MonadTrans m
-    => TermLike Variable
-    -> TermLike Variable
-    -> m Simplifier (Either (Doc ()) (NonEmpty (Predicate Variable)))
-runUnifier first second = do
-    unifier <- Lens.use lensUnifier
-    mvar <- Lens.use lensLogger
-    liftSimplifierWithLogger mvar
-        . runUnifierWithExplanation
-        $ unifier first second
-
-getNodeState :: InnerGraph -> Graph.Node -> Maybe (NodeState, Graph.Node)
-getNodeState graph node =
-        fmap (\nodeState -> (nodeState, node))
-        . strategyPattern StrategyPatternTransformer
-            { rewriteTransformer = const . Just $ UnevaluatedNode
-            , stuckTransformer = const . Just $ StuckNode
-            , bottomValue = Nothing
-            }
-        . Graph.lab'
-        . Graph.context graph
-        $ node
 
 data NodeState = StuckNode | UnevaluatedNode
     deriving (Eq, Ord, Show)
@@ -754,74 +475,3 @@ data GraphProofStatus
     | TrustedClaim
     deriving (Eq, Show)
 
--- | Adds or updates the provided alias.
-addOrUpdateAlias
-    :: forall m claim
-    .  MonadState (ReplState claim) m
-    => MonadError AliasError m
-    => AliasDefinition
-    -> m ()
-addOrUpdateAlias alias@AliasDefinition { name, command } = do
-    checkNameIsNotUsed
-    checkCommandExists
-    lensAliases Lens.%= Map.insert name alias
-  where
-    checkNameIsNotUsed :: m ()
-    checkNameIsNotUsed =
-        not . Set.member name <$> existingCommands
-            >>= falseToError NameAlreadyDefined
-
-    checkCommandExists :: m ()
-    checkCommandExists = do
-        cmds <- existingCommands
-        let
-            maybeCommand = listToMaybe $ words command
-            maybeExists = Set.member <$> maybeCommand <*> pure cmds
-        maybe
-            (Monad.Error.throwError UnknownCommand)
-            (falseToError UnknownCommand)
-            maybeExists
-
-    existingCommands :: m (Set String)
-    existingCommands =
-        Set.union commandSet
-        . Set.fromList
-        . Map.keys
-        <$> Lens.use lensAliases
-
-    falseToError :: AliasError -> Bool -> m ()
-    falseToError e b =
-        if b then pure () else Monad.Error.throwError e
-
-
-findAlias
-    :: MonadState (ReplState claim) m
-    => String
-    -> m (Maybe AliasDefinition)
-findAlias name = Map.lookup name <$> Lens.use lensAliases
-
-substituteAlias
-    :: AliasDefinition
-    -> ReplAlias
-    -> String
-substituteAlias
-    AliasDefinition { arguments, command }
-    ReplAlias { arguments = actualArguments } =
-    unwords
-      . fmap replaceArguments
-      . words
-      $ command
-  where
-    values :: Map String AliasArgument
-    values
-      | length arguments == length actualArguments
-        = Map.fromList $ zip arguments actualArguments
-      | otherwise = Map.empty
-
-    replaceArguments :: String -> String
-    replaceArguments s = maybe s toString $ Map.lookup s values
-
-    toString :: AliasArgument -> String
-    toString = \case
-        SimpleArgument str -> str
-        QuotedArgument str -> "\"" <> str <> "\""
