+++ kore/src/Kore/Repl/Interpreter.hs
@@ -51,13 +51,11 @@ import qualified Data.Functor.Foldable as Recursive
 import qualified Data.Graph.Inductive.Graph as Graph
 import qualified Data.GraphViz as Graph
 import qualified Data.List as List
-import           Data.List.Extra
-                 ( groupSort )
 import           Data.List.NonEmpty
                  ( NonEmpty )
 import qualified Data.Map.Strict as Map
 import           Data.Maybe
-                 ( catMaybes, listToMaybe )
+                 ( listToMaybe )
 import           Data.Sequence
                  ( Seq )
 import qualified Data.Text as Text
@@ -94,6 +92,7 @@ import           Kore.Repl.Data
 import           Kore.Repl.Parser
 import           Kore.Repl.Parser
                  ( commandParser )
+import           Kore.Repl.State
 import           Kore.Step.Rule
                  ( RewriteRule (..), RulePattern (..) )
 import qualified Kore.Step.Rule as Rule
@@ -211,6 +210,13 @@ exit
     => ReplM claim ReplStatus
 exit = do
     proofs <- allProofs
+    ofile <- Lens.use lensOutputFile
+    let fileName =
+            maybe (error "Output file not specified") id (unOutputFile ofile)
+    onePathClaims <- generateInProgressOPClaims
+    sort <- currentClaimSort
+    let conj = conjOfOnePathClaims onePathClaims sort
+    liftIO $ writeFile fileName (unparseToString conj)
     if isCompleted (Map.elems proofs)
        then return SuccessStop
        else return FailStop
@@ -1050,19 +1056,6 @@ formatUnificationMessage =
         . map (Pretty.indent 4 . unparseUnifier)
         . Foldable.toList
     unparseUnifier = unparse . (<$) (TermLike.mkTop_ :: TermLike Variable)
-
-findLeafNodes :: InnerGraph -> [Graph.Node]
-findLeafNodes graph =
-    filter ((==) 0 . Graph.outdeg graph) $ Graph.nodes graph
-
-sortLeafsByType :: InnerGraph -> Map.Map NodeState [Graph.Node]
-sortLeafsByType graph =
-    Map.fromList
-        . groupSort
-        . catMaybes
-        . fmap (getNodeState graph)
-        . findLeafNodes
-        $ graph
 
 showProofStatus :: Map.Map ClaimIndex GraphProofStatus -> String
 showProofStatus m =
