+++ kore/test/Test/Kore/Repl/Interpreter.hs
@@ -33,6 +33,7 @@ import           Kore.OnePath.Verification
                  ( Axiom (..), verifyClaimStep )
 import           Kore.Repl.Data
 import           Kore.Repl.Interpreter
+import           Kore.Repl.State
 import           Kore.Step.Rule
                  ( OnePathRule (..), RewriteRule (..), rulePattern )
 
@@ -473,6 +474,7 @@ mkState axioms claims claim logger =
         , aliases     = Map.empty
         , logging     = (Logger.Debug, NoLogging)
         , logger
+        , outputFile  = OutputFile Nothing
         }
   where
     graph' = emptyExecutionGraph claim
